Crate private_set_intersection

Private Set Intersection example for Project Oak.

Clients invoke the module by providing their own private set, and the module keeps track of the intersection of all the provided sets from all the clients that have interacted with it.

The (common) intersection can then be retrieved by each client by a separate invocation.

TODO(#747): Limit number of clients / sets that may contribute to the intersection. TODO(#747): Consider stopping accepting contributions after the first client retrieves the intersection.

Functions

oak_main